Move debug role change button to sidebar
